Fix HoverTest for screens with a camera notch
Use Instrumentation.sendPointerSync instead of Window.injectInputEvent
to adjust for the window position which can be not (0,0) when camera
notch is present.

Use the default theme for the test activity because the previously
used theme was fullscreen and this led to the test trying to click
over a status bar.

Bug: 72581119
Test: android.view.cts.HoverTest
Change-Id: I602bb2ad49bac3edfcf509c4a6317b8867c3cf40
(cherry picked from commit d63f527320c82612312f8d8e04527125f50e52b6)