Check all versions of system APIs that platform supports
Platform is now supposed to support "multiple" versions of System APIs,
otherwise any backward incompatible change in the System API has a risk
of breaking all existing clients are unbundled in the vendor partition.

The range of System API version that the platform supports is determined
by setting PLATFORM_SYSTEMSDK_MIN_VERSION and the range is shown as
PLATFORM_SYSTEMSDK_VERSIONS. Platform has to support all the APIs (not
just the latest one) in the range. CtsSystemCurrentApiSignatureTest is
ranamed to CtsSystemApiSignatureTest and modified to ensure this.

Specifically,

1) the API files for the old versions of System API are also
genererated.

2) they are zipped into a single file and unzipped in the test case and
parsed there. This zipping is required because the range of supported
System API versions will vary for each platform releases, while
the list of API files to be copied to the device are hard coded in the
AndroidTest.xml file.

Bug: 69900663
Test: atest under this directory
Test: copy frameworks/base/api/system-current.txt into
prebuilts/sdk/system-api/27.txt and manually add some APIs there.
Set PLATFORM_SYSTEMSDK_MIN_VERSION to 27 in build/make/core/verion_defaults.mk
atest under this directory. The manually added APIs are correctly identified as missing.

Change-Id: I237cf0ef73e2476df325cf0f7040414e8f315b82