[Magnifier-30] Add test for window content size
Since the size of the magnifier window was exposed in the public API,
this CL adds a check that the size of the magnifier window content
matches the size of the magnifier.

Also, the CL updates the way we test that the magnifier content contains
4 quadrants, in order to be less sensitive to noise. This noise is
likely to occur if the magnified view region is upscaled and filtered,
so it might represent a common case.

Bug: 69407609
Test: atest CtsWidgetTestCases:android.widget.cts.MagnifierTest
Change-Id: I96e29c48d24e5987bddcd14b59c86975431648f5
(cherry picked from commit 1459edc862b10bf666d818ee80ef065c1e88f475)
Merged-In: I96e29c48d24e5987bddcd14b59c86975431648f5