Fix the DeviceAndProfileOwner CTS test around VPN disallowed apps.
This was broken by a change that made it so that disallowed apps
won't be told any more by network callbacks about the VPNs they
are not allowed to see by policy.
Of course it makes sense that they can't, but this test was relying
on the running test (which is the app under test and is thus
disallowed in this test) getting the callback to know when the VPN
is finished bringing up.

This change fixes this by listening to the VPN in the VPN app
instead, and having the VPN app send an intent to the app under
test when the VPN is up.

Bug: 73085243
Test: Previously failing CTS now passes ; other affected CTS still passing
Change-Id: Id6bd8dc83ec4e523cf62c9d3264daa082754f117
Merged-In: I30715494204ce50313ee48ae84854e789484aaa2