Wait between switching user and dismissing the keyguard.
Otherwise, the test can become flaky (especially on GCE) due to the user
not having finished switching between the keyguard dismissal attempt.

Other cases of #switchUser don't try to dismiss the keyguard after
without significant work in-between, so this is the only one updated.

Documentation has been added to #switchUser to make this clear.

Bug: 74231690
Change-Id: Ibe6ec509ee2a2f19966a564bf8aac5fb9dfb363b
Fixes: 74231690
Test: Only test changes