DO NOT MERGE [CTS] Add a test for key/value backup and restore.
This test is intended to replace BackupTest in CTS Verifier (removed in
this change).

The test installs the app that generates random values and saves them in
different shared preferences and files. It uses BackupAgentHelper to do
key/value backup of those values. The tests verifies that the values are
restored after the app is uninstalled and reinstalled.

Test: make cts -j40 && cts-tradefed run cts -m CtsBackupHostTestCases

b/34127370

Change-Id: Ic2c5ca6244b2c66ec4596dd553284d3d62b9ce3d
Signed-off-by: Raj Mamadgi <r.mamadgi@samsung.com>