Make testUidTagStateDetails more tolerant.
testUidTagStateDetails checks various combinations of tag and
foreground/background state and expects them to return results
that are all within 10% of the first result.

Unfortunately, this causes the test to fail when run by itself.
This is because the test compares tagged traffic (~20kB, caused
by the test itself) with untagged traffic (which also includes
~5kB of traffic caused by the test harness). Fix this by
disabling compression so that the 20kB becomes 100kB.

Bug: 78218117
Test: NetworkUsageStatsTest passes on taimen
Change-Id: Ie4d909096564ff37ef03fbb53b8a374f4bc122c0