Remove erroneous calls made to a ListView object.
In the BleConnectionPriorityClientBaseActivity class the onCreate method
makes calls to setEnabled() and setClickable() on a ListView object.
These methods aren't seen anywhere in the online ListView documentation
or the ListView code itself. Additionally, the
BleEncryptedClientBaseActivity and BleClientTestBaseActivity classes,
which are in a similar test group, behave similarly but don't have the
erroneous calls.

Bug: 69674233
Test: Built CtsVerifier apk and installed on watch. No longer observed
issue.

Change-Id: Ieb27a5f7e8f3db6889f928dfe981e0a947dbceba