Trigger lockNow() via a broadcast
At the moment when we need to lock the work profile, we execute a dummy test
case which calls lockNow(). This leads to problems because lockNow() will
result in all work apps including the test process being killed, which is
interpreted by tradefed as a test failure "Process crashed". To fix this
issue, we switch to sending a broadcast to  invoke lockNow().

Bug: 93278874
Test: atest com.android.cts.devicepolicy.ManagedProfileTest#testResetPasswordWithTokenBeforeUnlock
Test: atest com.android.cts.devicepolicy.ManagedProfileTest#testResetPasswordTokenUsableAfterClearingLock
Test: atest com.android.cts.devicepolicy.ManagedProfileTest#testLockNowWithKeyEviction
Change-Id: Ib380fde4cb6376cacfd589a254398e1688868e54